remove: factory command and internal/factory package
The factory feature (supermodel factory health|run|improve) is removed.
The health analysis logic it contained is moved into a new internal/audit/
vertical slice so the existing 'supermodel audit' command continues to work.

- delete cmd/factory.go
- delete internal/factory/ (health.go, render.go, types.go, zip.go, tests)
- add internal/audit/ with the health analysis, rendering, and zip logic
- update cmd/audit.go to import internal/audit instead of internal/factory
