+++ internal/audit/zip.go
@@ -1,4 +1,4 @@
-package factory
+package audit
 
 import (
 	"archive/zip"
